Rewrite movement mechanism
From now on, each user input triggers a "turn". Each "turn" consists of
multiple "steps".

One step consists of movements that happen simultaneously, i.e. which
should be animated together.

In the long run this will allow us to yield information on each step,
which is needed for animating the movement.